File uploaded from jupyter notebook
The file "Youtube Adview Prediction.ipynb" seems to be a Jupyter Notebook file, typically used for interactive computing and data analysis tasks. The ".ipynb" extension stands for IPython Notebook, which is now commonly used with Jupyter Notebooks. Based on the name, it appears to be related to predicting YouTube ad views, likely through some form of machine learning or statistical analysis.